pyproject.toml
[project]
name = "carpriceprediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "catboost>=1.2.8",
    "dvc>=3.64.2",
    "matplotlib>=3.10.7",
    "mlflow<3.0",
    "numpy>=2.3.5",
    "optuna>=4.6.0",
    "pandantic>=1.0.1",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "xgboost>=3.1.2",
]
